nta.c: be more tolerant to trivial syntax errors
nta used to respond with 400 Bad Xyzzy Header even if the most harmless
headers was malformed.

Problem reported by Mike Jerris.

darcs-hash:20061214223851-65a35-f612d478e493206881ec7c90554bd48e760dcef2.gz

+++ libsofia-sip-ua/nta/nta_internal.h
@@ -136,9 +136,9 @@ struct nta_agent_s
 
 
   /** Request error mask */
-  unsigned short        sa_bad_req_mask;
+  unsigned              sa_bad_req_mask;
   /** Response error mask */
-  unsigned short        sa_bad_resp_mask;
+  unsigned              sa_bad_resp_mask;
 
   /** Maximum size of incoming messages */
   size_t                sa_maxsize;
